Rename Lexer namespace to Parsing and introduce ParsingException
