+++ SafeRapidPdf/Objects/PdfArray.cs
@@ -15,7 +15,7 @@ namespace SafeRapidPdf.Objects
 
         public override IReadOnlyList<IPdfObject> Items => _items;
 
-        public static PdfArray Parse(Lexical.ILexer lexer)
+        public static PdfArray Parse(Parsing.ILexer lexer)
         {
             var list = new List<IPdfObject>();
             PdfObject value;
