+++ SafeRapidPdf/Objects/PdfObject.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Text;
 
+using SafeRapidPdf.Parsing;
+
 namespace SafeRapidPdf.Objects
 {
     public abstract class PdfObject : IPdfObject
@@ -17,12 +19,12 @@ namespace SafeRapidPdf.Objects
 
         public string Text => ToString();
 
-        public static PdfObject ParseAny(Lexical.ILexer lexer)
+        public static PdfObject ParseAny(ILexer lexer)
         {
             return ParseAny(lexer, string.Empty);
         }
 
-        public static PdfObject ParseAny(Lexical.ILexer lexer, string endToken)
+        public static PdfObject ParseAny(ILexer lexer, string endToken)
         {
             string token = lexer.ReadToken();
             if (token == null)
@@ -70,7 +72,10 @@ namespace SafeRapidPdf.Objects
                     // check for stream and combine put dictionary into stream object
                     token = lexer.PeekToken1();
                     if (token == "stream")
+                    {
                         obj = PdfStream.Parse(obj as PdfDictionary, lexer);
+                    }
+
                     break;
 
                 case "[":
@@ -90,8 +95,11 @@ namespace SafeRapidPdf.Objects
                 case "endstream":
                 case "endobj":
                     if (endToken == token)
+                    {
                         return null; // expected end
-                    throw new Exception("Parser error: out of sync");
+                    }
+
+                    throw new ParsingException("Out of sync");
 
                 default:
                     // must be an integer or double value
@@ -122,8 +130,12 @@ namespace SafeRapidPdf.Objects
                     }
                     break;
             }
+
             if (obj == null)
-                throw new Exception("Parse error, could not read object");
+            {
+                throw new ParsingException("Could not read object");
+            }
+
             return obj;
         }
