+++ SafeRapidPdf/Objects/PdfTrailer.cs
@@ -7,7 +7,7 @@
         {
         }
 
-        public static new PdfTrailer Parse(Lexical.ILexer lexer)
+        public static new PdfTrailer Parse(Parsing.ILexer lexer)
         {
             lexer.Expects("<<");
             var dictionary = PdfDictionary.Parse(lexer);
