+++ SafeRapidPdf/Parsing/ParsingException.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace SafeRapidPdf.Parsing
+{
+    public class ParsingException : Exception
+    {
+        public ParsingException(string message)
+            : base(message) { }
+    }
+}
